@@ -1,3 +1,3 @@
 module github.com/anishathalye/porcupine
 
-go 1.16
+go 1.18
