@@ -3,7 +3,7 @@ package porcupine
 import "time"
 
 // CheckOperations checks whether a history is linearizable.
-func CheckOperations(model Model, history []Operation) bool {
+func CheckOperations[S, I, O any](model Model[S, I, O], history []Operation) bool {
 	res, _ := checkOperations(model, history, false, 0)
 	return res == Ok
 }
@@ -12,7 +12,7 @@ func CheckOperations(model Model, history []Operation) bool {
 // timeout.
 //
 // A timeout of 0 is interpreted as an unlimited timeout.
-func CheckOperationsTimeout(model Model, history []Operation, timeout time.Duration) CheckResult {
+func CheckOperationsTimeout[S, I, O any](model Model[S, I, O], history []Operation, timeout time.Duration) CheckResult {
 	res, _ := checkOperations(model, history, false, timeout)
 	return res
 }
@@ -21,12 +21,12 @@ func CheckOperationsTimeout(model Model, history []Operation, timeout time.Durat
 // computing data that can be used to visualize the history and linearization.
 //
 // The returned LinearizationInfo can be used with [Visualize].
-func CheckOperationsVerbose(model Model, history []Operation, timeout time.Duration) (CheckResult, LinearizationInfo) {
+func CheckOperationsVerbose[S, I, O any](model Model[S, I, O], history []Operation, timeout time.Duration) (CheckResult, LinearizationInfo) {
 	return checkOperations(model, history, true, timeout)
 }
 
 // CheckEvents checks whether a history is linearizable.
-func CheckEvents(model Model, history []Event) bool {
+func CheckEvents[S, I, O any](model Model[S, I, O], history []Event) bool {
 	res, _ := checkEvents(model, history, false, 0)
 	return res == Ok
 }
@@ -34,7 +34,7 @@ func CheckEvents(model Model, history []Event) bool {
 // CheckEventsTimeout checks whether a history is linearizable, with a timeout.
 //
 // A timeout of 0 is interpreted as an unlimited timeout.
-func CheckEventsTimeout(model Model, history []Event, timeout time.Duration) CheckResult {
+func CheckEventsTimeout[S, I, O any](model Model[S, I, O], history []Event, timeout time.Duration) CheckResult {
 	res, _ := checkEvents(model, history, false, timeout)
 	return res
 }
@@ -43,6 +43,6 @@ func CheckEventsTimeout(model Model, history []Event, timeout time.Duration) Che
 // data that can be used to visualize the history and linearization.
 //
 // The returned LinearizationInfo can be used with [Visualize].
-func CheckEventsVerbose(model Model, history []Event, timeout time.Duration) (CheckResult, LinearizationInfo) {
+func CheckEventsVerbose[S, I, O any](model Model[S, I, O], history []Event, timeout time.Duration) (CheckResult, LinearizationInfo) {
 	return checkEvents(model, history, true, timeout)
 }
